Found the max, I guess

+++ project_3/part1/array_gen.c
@@ -1,32 +1,32 @@
 #include "array_gen.h"
 
 
-double * make_array(int SIZE) {
+double * make_array(long int SIZE) {
 	double * array = (double *) malloc(SIZE * sizeof(double));
 	srand(time(NULL));
 
-	int i;
+	long int i;
 	for (i = 0; i < SIZE; i++) {
 		array[i] = (double)rand() / (double)RAND_MAX * 1000;
 	}
 	return array;
 }
 
-void print_array(double * array, int SIZE) {
-	int i;
+void print_array(double * array, long int SIZE) {
+	long int i;
 	for (i = (SIZE - 25); i < SIZE; i++) {
 		printf("%f\n", array[i]);
 	}
 }
 
-void write_array(double * array, char * filename, int SIZE) {
+void write_array(double * array, char * filename, long int SIZE) {
 	FILE *f = fopen(filename, "wb");
 	printf("Writing Array\n");
 	fwrite(array, sizeof(double), SIZE, f);
 	fclose(f);
 }
 
-double * read_array(char * filename, int SIZE) {
+double * read_array(char * filename, long int SIZE) {
 	double * array = (double *) malloc(SIZE * sizeof(double));
 	FILE *f = fopen(filename, "rb");
 	if (f == NULL) {
