+++ project_3/part1/array_gen.h
@@ -5,9 +5,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-double * make_array(int SIZE);
-void print_array(double * array, int SIZE);
-void write_array(double * array, char * filename, int SIZE);
-double * read_array(char * filename, int SIZE);
+double * make_array(long int SIZE);
+void print_array(double * array, long int SIZE);
+void write_array(double * array, char * filename, long int SIZE);
+double * read_array(char * filename, long int SIZE);
 
 #endif
